statistic service added

+++ src/main/java/com/getir/casestudy/model/dto/BookDetailDTO.java
@@ -12,5 +12,5 @@ public class BookDetailDTO {
 
     @NotNull(message = "Book id count required")
     @Min(value = 1, message = "Book count must greater than zero")
-    private Long bookCount;
+    private Integer bookCount;
 }
